Alternative Impl. to kqueue.

+++ pluto_os_abstraction/src/system_events/pluto_system_events.c
@@ -360,7 +360,7 @@ int32_t PLUTO_SystemEventsPoll(PLUTO_SystemEventHandler_t handler, PLUTO_SystemE
         handler->epoll.epoll_fd,
         handler->inotify.inotify_fd
     };
-    for(int i=0;i<(sizeof(filedescriptors)/sizeof(int));++i)
+    for(size_t i=0;i<(sizeof(filedescriptors)/sizeof(int));++i)
     {
         int res = epoll_wait(filedescriptors[i], events, PLUTO_MAX_EPOLL_EVENTS_BUFFER, 0);
         if(res < 0)
